The functions for creating data sets with predetermined 1-3 split levels have been added to the ‘create_testdata’ file.
create_X, create_K, create_T and create_Y each generate a data set with a predetermined split in variables X, Xk, Xt and Y respectively, whereby the correlation between X and Y is always increased. Xk and Xt are third variables, where Xk is an ordinal/metric variable and Xt is a nominal variable.

The strength of the target correlation and the increase in correlation per split can be set.

The following functions generate data sets that are designed for one or two more split levels (two or three in total). The procedure remains the same; however, the variables that ideally split the data set are always different in pairs within a data set. The order of the split variables was taken into account. (For var1 = Xk; var2 = Y, it makes more sense to split first into Xk, then into Y than the other way round).

The functions that allow data sets with all four split levels (X, Xk, Xt, Y) in all possible permutations with regard to the split order are still pending.