The argument ‘type’ with possible values ‘rel’ and ‘abs’ has been added to the function ‘calc_sampi’.
By entering ‘rel’ (default), the relative sampi value is determined, i.e. only nodes that lead to a split in the ‘true_tree’ are included in the calculation.

If ‘abs’ is entered, however, those nodes that lead to a split in ‘test_tree’ but not in ‘true_tree’ are also taken into account.

In both cases, a split that is only present in one of the two trees allways leads to a higher sampi value. The absolute sampi value is therefore always greater than or equal to the relative sampi value.